+++ src/main/java/com/github/michaeldoyle/istat4j/model/TelemetryCpu.java
@@ -1,5 +1,5 @@
 /* 
- * Copyright 2013 Michael J Doyle
+ * Copyright 2013-2014 Michael J Doyle
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,18 +15,20 @@
  */
 package com.github.michaeldoyle.istat4j.model;
 
+import java.math.BigInteger;
+
 public class TelemetryCpu {
 
-	private long id;
+	private BigInteger id;
 	private int user;
 	private int system;
 	private int nice;
 	
-	public long getId() {
+	public BigInteger getId() {
 		return id;
 	}
 	
-	public void setId(long id) {
+	public void setId(BigInteger id) {
 		this.id = id;
 	}
